Fix: Adapted to more organized MVVM to avoid merge conflicts
