+++ app/src/main/AndroidManifest.xml
@@ -13,24 +13,24 @@
         android:theme="@style/Theme.SportHUB"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.home.MainActivity"
             android:exported="true" />
         <activity
-            android:name=".GenderSelectionActivity"
+            android:name=".ui.login.GenderSelectionActivity"
             android:exported="false" />
         <activity
-            android:name=".BirthDateSelectionActivity"
+            android:name=".ui.login.BirthDateSelectionActivity"
             android:exported="false" />
         <activity
-            android:name=".NameSelectionActivity"
+            android:name=".ui.login.NameSelectionActivity"
             android:exported="false"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
         <activity
-            android:name=".FavoriteSportsSelectionActivity"
+            android:name=".ui.login.FavoriteSportsSelectionActivity"
             android:exported="false"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
         <activity
-            android:name=".SignInActivity"
+            android:name=".ui.login.SignInActivity"
             android:exported="true"
             android:label="@string/app_name">
             <intent-filter>
