+++ app/src/main/java/com/example/sporthub/data/model/Sport.kt
@@ -1,0 +1,8 @@
+// com.example.sporthub.data.model.Sport.kt
+package com.example.sporthub.data.model
+
+data class Sport(
+    val id: String,
+    val name: String,
+    val logo: String
+)
