--- app/src/main/java/com/example/sporthub/ui/login/LoginActivity.kt
@@ -1,4 +1,0 @@
-package com.example.sporthub.ui.login
-
-
-
