+++ app/src/main/java/com/example/sporthub/ui/login/SignInActivity.kt
@@ -1,4 +1,4 @@
-package com.example.sporthub
+package com.example.sporthub.ui.login
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
@@ -6,6 +6,8 @@ import android.os.Bundle
 import android.util.Log
 import android.widget.Button
 import android.widget.Toast
+import com.example.sporthub.R
+import com.example.sporthub.ui.home.MainActivity
 import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import com.google.android.gms.common.api.ApiException
