+++ app/src/main/res/layout/activity_birth_date_selection.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:paddingTop="24dp"
-    tools:context=".BirthDateSelectionActivity">
+    tools:context=".ui.login.BirthDateSelectionActivity">
 
     <TextView
         android:id="@+id/textview_title"
