+++ app/src/main/res/layout/activity_favorite_sports_selection.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".FavoriteSportsSelectionActivity">
+    tools:context=".ui.login.FavoriteSportsSelectionActivity">
 
     <TextView
         android:id="@+id/textview_title"
