+++ app/src/main/res/layout/activity_gender_selection.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".GenderSelectionActivity">
+    tools:context=".ui.login.GenderSelectionActivity">
 
     <TextView
         android:id="@+id/textview_title"
