+++ app/src/main/res/layout/activity_name_selection.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".NameSelectionActivity">
+    tools:context=".ui.login.NameSelectionActivity">
 
     <TextView
         android:id="@+id/textview_title"
